Convert to Rollup.js and ES6
