+++ .gitignore
@@ -4,3 +4,4 @@
 node_modules
 npm-debug.log
 yarn.lock
+test/utils/timezones.js
