--- jstz.rules.js
@@ -1,917 +1,0 @@
-/* Build time: 2015-11-02 13:01:00Z Build by invoking python utilities/dst.py generate */
-jstz.olson.dst_rules = {
-    "years": [
-        2008,
-        2009,
-        2010,
-        2011,
-        2012,
-        2013,
-        2014
-    ],
-    "zones": [
-        {
-            "name": "Africa/Cairo",
-            "rules": [
-                {
-                    "e": 1219957200000,
-                    "s": 1209074400000
-                },
-                {
-                    "e": 1250802000000,
-                    "s": 1240524000000
-                },
-                {
-                    "e": 1285880400000,
-                    "s": 1284069600000
-                },
-                false,
-                false,
-                false,
-                {
-                    "e": 1411678800000,
-                    "s": 1406844000000
-                }
-            ]
-        },
-        {
-            "name": "Africa/Casablanca",
-            "rules": [
-                {
-                    "e": 1220223600000,
-                    "s": 1212278400000
-                },
-                {
-                    "e": 1250809200000,
-                    "s": 1243814400000
-                },
-                {
-                    "e": 1281222000000,
-                    "s": 1272758400000
-                },
-                {
-                    "e": 1312066800000,
-                    "s": 1301788800000
-                },
-                {
-                    "e": 1348970400000,
-                    "s": 1345428000000
-                },
-                {
-                    "e": 1382839200000,
-                    "s": 1376100000000
-                },
-                {
-                    "e": 1414288800000,
-                    "s": 1406944800000
-                }
-            ]
-        },
-        {
-            "name": "America/Asuncion",
-            "rules": [
-                {
-                    "e": 1205031600000,
-                    "s": 1224388800000
-                },
-                {
-                    "e": 1236481200000,
-                    "s": 1255838400000
-                },
-                {
-                    "e": 1270954800000,
-                    "s": 1286078400000
-                },
-                {
-                    "e": 1302404400000,
-                    "s": 1317528000000
-                },
-                {
-                    "e": 1333854000000,
-                    "s": 1349582400000
-                },
-                {
-                    "e": 1364094000000,
-                    "s": 1381032000000
-                },
-                {
-                    "e": 1395543600000,
-                    "s": 1412481600000
-                }
-            ]
-        },
-        {
-            "name": "America/Campo_Grande",
-            "rules": [
-                {
-                    "e": 1203217200000,
-                    "s": 1224388800000
-                },
-                {
-                    "e": 1234666800000,
-                    "s": 1255838400000
-                },
-                {
-                    "e": 1266721200000,
-                    "s": 1287288000000
-                },
-                {
-                    "e": 1298170800000,
-                    "s": 1318737600000
-                },
-                {
-                    "e": 1330225200000,
-                    "s": 1350792000000
-                },
-                {
-                    "e": 1361070000000,
-                    "s": 1382241600000
-                },
-                {
-                    "e": 1392519600000,
-                    "s": 1413691200000
-                }
-            ]
-        },
-        {
-            "name": "America/Goose_Bay",
-            "rules": [
-                {
-                    "e": 1225594860000,
-                    "s": 1205035260000
-                },
-                {
-                    "e": 1257044460000,
-                    "s": 1236484860000
-                },
-                {
-                    "e": 1289098860000,
-                    "s": 1268539260000
-                },
-                {
-                    "e": 1320555600000,
-                    "s": 1299988860000
-                },
-                {
-                    "e": 1352005200000,
-                    "s": 1331445600000
-                },
-                {
-                    "e": 1383454800000,
-                    "s": 1362895200000
-                },
-                {
-                    "e": 1414904400000,
-                    "s": 1394344800000
-                }
-            ]
-        },
-        {
-            "name": "America/Havana",
-            "rules": [
-                {
-                    "e": 1224997200000,
-                    "s": 1205643600000
-                },
-                {
-                    "e": 1256446800000,
-                    "s": 1236488400000
-                },
-                {
-                    "e": 1288501200000,
-                    "s": 1268542800000
-                },
-                {
-                    "e": 1321160400000,
-                    "s": 1300597200000
-                },
-                {
-                    "e": 1352005200000,
-                    "s": 1333256400000
-                },
-                {
-                    "e": 1383454800000,
-                    "s": 1362891600000
-                },
-                {
-                    "e": 1414904400000,
-                    "s": 1394341200000
-                }
-            ]
-        },
-        {
-            "name": "America/Mazatlan",
-            "rules": [
-                {
-                    "e": 1225008000000,
-                    "s": 1207472400000
-                },
-                {
-                    "e": 1256457600000,
-                    "s": 1238922000000
-                },
-                {
-                    "e": 1288512000000,
-                    "s": 1270371600000
-                },
-                {
-                    "e": 1319961600000,
-                    "s": 1301821200000
-                },
-                {
-                    "e": 1351411200000,
-                    "s": 1333270800000
-                },
-                {
-                    "e": 1382860800000,
-                    "s": 1365325200000
-                },
-                {
-                    "e": 1414310400000,
-                    "s": 1396774800000
-                }
-            ]
-        },
-        {
-            "name": "America/Mexico_City",
-            "rules": [
-                {
-                    "e": 1225004400000,
-                    "s": 1207468800000
-                },
-                {
-                    "e": 1256454000000,
-                    "s": 1238918400000
-                },
-                {
-                    "e": 1288508400000,
-                    "s": 1270368000000
-                },
-                {
-                    "e": 1319958000000,
-                    "s": 1301817600000
-                },
-                {
-                    "e": 1351407600000,
-                    "s": 1333267200000
-                },
-                {
-                    "e": 1382857200000,
-                    "s": 1365321600000
-                },
-                {
-                    "e": 1414306800000,
-                    "s": 1396771200000
-                }
-            ]
-        },
-        {
-            "name": "America/Miquelon",
-            "rules": [
-                {
-                    "e": 1225598400000,
-                    "s": 1205038800000
-                },
-                {
-                    "e": 1257048000000,
-                    "s": 1236488400000
-                },
-                {
-                    "e": 1289102400000,
-                    "s": 1268542800000
-                },
-                {
-                    "e": 1320552000000,
-                    "s": 1299992400000
-                },
-                {
-                    "e": 1352001600000,
-                    "s": 1331442000000
-                },
-                {
-                    "e": 1383451200000,
-                    "s": 1362891600000
-                },
-                {
-                    "e": 1414900800000,
-                    "s": 1394341200000
-                }
-            ]
-        },
-        {
-            "name": "America/Santa_Isabel",
-            "rules": [
-                {
-                    "e": 1225011600000,
-                    "s": 1207476000000
-                },
-                {
-                    "e": 1256461200000,
-                    "s": 1238925600000
-                },
-                {
-                    "e": 1288515600000,
-                    "s": 1270375200000
-                },
-                {
-                    "e": 1319965200000,
-                    "s": 1301824800000
-                },
-                {
-                    "e": 1351414800000,
-                    "s": 1333274400000
-                },
-                {
-                    "e": 1382864400000,
-                    "s": 1365328800000
-                },
-                {
-                    "e": 1414314000000,
-                    "s": 1396778400000
-                }
-            ]
-        },
-        {
-            "name": "America/Santiago",
-            "rules": [
-                {
-                    "e": 1206846000000,
-                    "s": 1223784000000
-                },
-                {
-                    "e": 1237086000000,
-                    "s": 1255233600000
-                },
-                {
-                    "e": 1270350000000,
-                    "s": 1286683200000
-                },
-                {
-                    "e": 1304823600000,
-                    "s": 1313899200000
-                },
-                {
-                    "e": 1335668400000,
-                    "s": 1346558400000
-                },
-                {
-                    "e": 1367118000000,
-                    "s": 1378612800000
-                },
-                {
-                    "e": 1398567600000,
-                    "s": 1410062400000
-                }
-            ]
-        },
-        {
-            "name": "America/Sao_Paulo",
-            "rules": [
-                {
-                    "e": 1203213600000,
-                    "s": 1224385200000
-                },
-                {
-                    "e": 1234663200000,
-                    "s": 1255834800000
-                },
-                {
-                    "e": 1266717600000,
-                    "s": 1287284400000
-                },
-                {
-                    "e": 1298167200000,
-                    "s": 1318734000000
-                },
-                {
-                    "e": 1330221600000,
-                    "s": 1350788400000
-                },
-                {
-                    "e": 1361066400000,
-                    "s": 1382238000000
-                },
-                {
-                    "e": 1392516000000,
-                    "s": 1413687600000
-                }
-            ]
-        },
-        {
-            "name": "Asia/Amman",
-            "rules": [
-                {
-                    "e": 1225404000000,
-                    "s": 1206655200000
-                },
-                {
-                    "e": 1256853600000,
-                    "s": 1238104800000
-                },
-                {
-                    "e": 1288303200000,
-                    "s": 1269554400000
-                },
-                {
-                    "e": 1319752800000,
-                    "s": 1301608800000
-                },
-                false,
-                false,
-                {
-                    "e": 1414706400000,
-                    "s": 1395957600000
-                }
-            ]
-        },
-        {
-            "name": "Asia/Damascus",
-            "rules": [
-                {
-                    "e": 1225486800000,
-                    "s": 1207260000000
-                },
-                {
-                    "e": 1256850000000,
-                    "s": 1238104800000
-                },
-                {
-                    "e": 1288299600000,
-                    "s": 1270159200000
-                },
-                {
-                    "e": 1319749200000,
-                    "s": 1301608800000
-                },
-                {
-                    "e": 1351198800000,
-                    "s": 1333058400000
-                },
-                {
-                    "e": 1382648400000,
-                    "s": 1364508000000
-                },
-                {
-                    "e": 1414702800000,
-                    "s": 1395957600000
-                }
-            ]
-        },
-        {
-            "name": "Asia/Dubai",
-            "rules": [
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Gaza",
-            "rules": [
-                {
-                    "e": 1219957200000,
-                    "s": 1206655200000
-                },
-                {
-                    "e": 1252015200000,
-                    "s": 1238104800000
-                },
-                {
-                    "e": 1281474000000,
-                    "s": 1269640860000
-                },
-                {
-                    "e": 1312146000000,
-                    "s": 1301608860000
-                },
-                {
-                    "e": 1348178400000,
-                    "s": 1333058400000
-                },
-                {
-                    "e": 1380229200000,
-                    "s": 1364508000000
-                },
-                {
-                    "e": 1414098000000,
-                    "s": 1395957600000
-                }
-            ]
-        },
-        {
-            "name": "Asia/Irkutsk",
-            "rules": [
-                {
-                    "e": 1224957600000,
-                    "s": 1206813600000
-                },
-                {
-                    "e": 1256407200000,
-                    "s": 1238263200000
-                },
-                {
-                    "e": 1288461600000,
-                    "s": 1269712800000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Jerusalem",
-            "rules": [
-                {
-                    "e": 1223161200000,
-                    "s": 1206662400000
-                },
-                {
-                    "e": 1254006000000,
-                    "s": 1238112000000
-                },
-                {
-                    "e": 1284246000000,
-                    "s": 1269561600000
-                },
-                {
-                    "e": 1317510000000,
-                    "s": 1301616000000
-                },
-                {
-                    "e": 1348354800000,
-                    "s": 1333065600000
-                },
-                {
-                    "e": 1382828400000,
-                    "s": 1364515200000
-                },
-                {
-                    "e": 1414278000000,
-                    "s": 1395964800000
-                }
-            ]
-        },
-        {
-            "name": "Asia/Kamchatka",
-            "rules": [
-                {
-                    "e": 1224943200000,
-                    "s": 1206799200000
-                },
-                {
-                    "e": 1256392800000,
-                    "s": 1238248800000
-                },
-                {
-                    "e": 1288450800000,
-                    "s": 1269698400000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Krasnoyarsk",
-            "rules": [
-                {
-                    "e": 1224961200000,
-                    "s": 1206817200000
-                },
-                {
-                    "e": 1256410800000,
-                    "s": 1238266800000
-                },
-                {
-                    "e": 1288465200000,
-                    "s": 1269716400000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Omsk",
-            "rules": [
-                {
-                    "e": 1224964800000,
-                    "s": 1206820800000
-                },
-                {
-                    "e": 1256414400000,
-                    "s": 1238270400000
-                },
-                {
-                    "e": 1288468800000,
-                    "s": 1269720000000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Vladivostok",
-            "rules": [
-                {
-                    "e": 1224950400000,
-                    "s": 1206806400000
-                },
-                {
-                    "e": 1256400000000,
-                    "s": 1238256000000
-                },
-                {
-                    "e": 1288454400000,
-                    "s": 1269705600000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Yakutsk",
-            "rules": [
-                {
-                    "e": 1224954000000,
-                    "s": 1206810000000
-                },
-                {
-                    "e": 1256403600000,
-                    "s": 1238259600000
-                },
-                {
-                    "e": 1288458000000,
-                    "s": 1269709200000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Yekaterinburg",
-            "rules": [
-                {
-                    "e": 1224968400000,
-                    "s": 1206824400000
-                },
-                {
-                    "e": 1256418000000,
-                    "s": 1238274000000
-                },
-                {
-                    "e": 1288472400000,
-                    "s": 1269723600000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Asia/Yerevan",
-            "rules": [
-                {
-                    "e": 1224972000000,
-                    "s": 1206828000000
-                },
-                {
-                    "e": 1256421600000,
-                    "s": 1238277600000
-                },
-                {
-                    "e": 1288476000000,
-                    "s": 1269727200000
-                },
-                {
-                    "e": 1319925600000,
-                    "s": 1301176800000
-                },
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Australia/Lord_Howe",
-            "rules": [
-                {
-                    "e": 1207407600000,
-                    "s": 1223134200000
-                },
-                {
-                    "e": 1238857200000,
-                    "s": 1254583800000
-                },
-                {
-                    "e": 1270306800000,
-                    "s": 1286033400000
-                },
-                {
-                    "e": 1301756400000,
-                    "s": 1317483000000
-                },
-                {
-                    "e": 1333206000000,
-                    "s": 1349537400000
-                },
-                {
-                    "e": 1365260400000,
-                    "s": 1380987000000
-                },
-                {
-                    "e": 1396710000000,
-                    "s": 1412436600000
-                }
-            ]
-        },
-        {
-            "name": "Australia/Perth",
-            "rules": [
-                {
-                    "e": 1206813600000,
-                    "s": 1224957600000
-                },
-                false,
-                false,
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Europe/Helsinki",
-            "rules": [
-                {
-                    "e": 1224982800000,
-                    "s": 1206838800000
-                },
-                {
-                    "e": 1256432400000,
-                    "s": 1238288400000
-                },
-                {
-                    "e": 1288486800000,
-                    "s": 1269738000000
-                },
-                {
-                    "e": 1319936400000,
-                    "s": 1301187600000
-                },
-                {
-                    "e": 1351386000000,
-                    "s": 1332637200000
-                },
-                {
-                    "e": 1382835600000,
-                    "s": 1364691600000
-                },
-                {
-                    "e": 1414285200000,
-                    "s": 1396141200000
-                }
-            ]
-        },
-        {
-            "name": "Europe/Minsk",
-            "rules": [
-                {
-                    "e": 1224979200000,
-                    "s": 1206835200000
-                },
-                {
-                    "e": 1256428800000,
-                    "s": 1238284800000
-                },
-                {
-                    "e": 1288483200000,
-                    "s": 1269734400000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Europe/Moscow",
-            "rules": [
-                {
-                    "e": 1224975600000,
-                    "s": 1206831600000
-                },
-                {
-                    "e": 1256425200000,
-                    "s": 1238281200000
-                },
-                {
-                    "e": 1288479600000,
-                    "s": 1269730800000
-                },
-                false,
-                false,
-                false,
-                false
-            ]
-        },
-        {
-            "name": "Pacific/Apia",
-            "rules": [
-                false,
-                false,
-                false,
-                {
-                    "e": 1301752800000,
-                    "s": 1316872800000
-                },
-                {
-                    "e": 1333202400000,
-                    "s": 1348927200000
-                },
-                {
-                    "e": 1365256800000,
-                    "s": 1380376800000
-                },
-                {
-                    "e": 1396706400000,
-                    "s": 1411826400000
-                }
-            ]
-        },
-        {
-            "name": "Pacific/Fiji",
-            "rules": [
-                false,
-                false,
-                {
-                    "e": 1269698400000,
-                    "s": 1287842400000
-                },
-                {
-                    "e": 1327154400000,
-                    "s": 1319292000000
-                },
-                {
-                    "e": 1358604000000,
-                    "s": 1350741600000
-                },
-                {
-                    "e": 1390050000000,
-                    "s": 1382796000000
-                },
-                {
-                    "e": 1421503200000,
-                    "s": 1414850400000
-                }
-            ]
-        },
-        {
-            "name": "Europe/London",
-            "rules": [
-                {
-                    "e": 1224982800000,
-                    "s": 1206838800000
-                },
-                {
-                    "e": 1256432400000,
-                    "s": 1238288400000
-                },
-                {
-                    "e": 1288486800000,
-                    "s": 1269738000000
-                },
-                {
-                    "e": 1319936400000,
-                    "s": 1301187600000
-                },
-                {
-                    "e": 1351386000000,
-                    "s": 1332637200000
-                },
-                {
-                    "e": 1382835600000,
-                    "s": 1364691600000
-                },
-                {
-                    "e": 1414285200000,
-                    "s": 1396141200000
-                }
-            ]
-        }
-    ]
-};
