+++ package.json
@@ -1,12 +1,16 @@
 {
   "name": "jstz",
   "version": "2.1.1",
-  "main": "index.js",
+  "main": "dist/jstz.min.js",
+  "module": "dist/jstz.esm.js",
   "description": "Timezone detection for JavaScript",
+  "license": "MIT",
+  "files": ["LICENSE", "README.md", "dist"],
   "scripts": {
-    "build": "gulp build",
+    "prebuild": "python ./utilities/dst.py generate",
+    "build": "rollup -c",
     "pretest": "npm run build",
-    "test": "./node_modules/.bin/mocha",
+    "test": "mocha",
     "bump:patch": "npm run build && npm version patch -m \"v%s\"",
     "bump:minor": "npm run build && npm version minor -m \"v%s\"",
     "bump:major": "npm run build && npm version major -m \"v%s\"",
@@ -42,18 +46,27 @@
     }
   ],
   "devDependencies": {
-    "chai": "^3.2.0",
+    "@babel/core": "^7.2.2",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "^7.2.3",
+    "chai": "^4.2.0",
     "del": "^0.1.3",
-    "gulp": "^3.8.10",
-    "gulp-concat": "^2.4.2",
-    "gulp-footer": "^1.0.5",
-    "gulp-header": "^1.2.2",
-    "gulp-jshint": "^1.9.0",
-    "gulp-rename": "^1.2.0",
-    "gulp-uglify": "^1.0.1",
-    "gulp-wrap-commonjs": "^0.1.12",
-    "mocha": "^2.3.0"
+    "eslint": "^5.11.1",
+    "mocha": "^5.2.0",
+    "rollup": "^1.0.0",
+    "rollup-plugin-babel": "^4.1.0",
+    "rollup-plugin-commonjs": "^9.2.0",
+    "rollup-plugin-conditional": "^2.2.0",
+    "rollup-plugin-node-resolve": "^4.0.0",
+    "rollup-plugin-terser": "^3.0.0",
+    "terser": "^3.13.1"
   },
   "typings": "./dist/jstz.d.ts",
-  "types": "./dist/jstz.d.ts"
+  "types": "./dist/jstz.d.ts",
+  "browserslist": [
+    "last 2 version",
+    "> 1%",
+    "not dead",
+    "not ie <= 9"
+  ]
 }
