+++ src/jstz.rules.js
@@ -1,0 +1,642 @@
+/* Build time: 2018-12-29 09:32:00Z Build by invoking python utilities/dst.py generate */
+export default {
+    "years": [
+        2014,
+        2015,
+        2016,
+        2017,
+        2018
+    ],
+    "zones": [
+        {
+            "name": "Africa/Cairo",
+            "rules": [
+                {
+                    "e": 1411678800000,
+                    "s": 1406844000000
+                },
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Africa/Casablanca",
+            "rules": [
+                {
+                    "e": 1414288800000,
+                    "s": 1406944800000
+                },
+                {
+                    "e": 1445738400000,
+                    "s": 1437271200000
+                },
+                {
+                    "e": 1477792800000,
+                    "s": 1468116000000
+                },
+                {
+                    "e": 1509242400000,
+                    "s": 1498960800000
+                },
+                {
+                    "e": 1540692000000,
+                    "s": 1529200800000
+                }
+            ]
+        },
+        {
+            "name": "America/Asuncion",
+            "rules": [
+                {
+                    "e": 1395543600000,
+                    "s": 1412481600000
+                },
+                {
+                    "e": 1426993200000,
+                    "s": 1443931200000
+                },
+                {
+                    "e": 1459047600000,
+                    "s": 1475380800000
+                },
+                {
+                    "e": 1490497200000,
+                    "s": 1506830400000
+                },
+                {
+                    "e": 1521946800000,
+                    "s": 1538884800000
+                }
+            ]
+        },
+        {
+            "name": "America/Campo_Grande",
+            "rules": [
+                {
+                    "e": 1392519600000,
+                    "s": 1413691200000
+                },
+                {
+                    "e": 1424574000000,
+                    "s": 1445140800000
+                },
+                {
+                    "e": 1456023600000,
+                    "s": 1476590400000
+                },
+                {
+                    "e": 1487473200000,
+                    "s": 1508040000000
+                },
+                {
+                    "e": 1518922800000,
+                    "s": 1541304000000
+                }
+            ]
+        },
+        {
+            "name": "America/Goose_Bay",
+            "rules": [
+                {
+                    "e": 1414904400000,
+                    "s": 1394344800000
+                },
+                {
+                    "e": 1446354000000,
+                    "s": 1425794400000
+                },
+                {
+                    "e": 1478408400000,
+                    "s": 1457848800000
+                },
+                {
+                    "e": 1509858000000,
+                    "s": 1489298400000
+                },
+                {
+                    "e": 1541307600000,
+                    "s": 1520748000000
+                }
+            ]
+        },
+        {
+            "name": "America/Havana",
+            "rules": [
+                {
+                    "e": 1414904400000,
+                    "s": 1394341200000
+                },
+                {
+                    "e": 1446354000000,
+                    "s": 1425790800000
+                },
+                {
+                    "e": 1478408400000,
+                    "s": 1457845200000
+                },
+                {
+                    "e": 1509858000000,
+                    "s": 1489294800000
+                },
+                {
+                    "e": 1541307600000,
+                    "s": 1520744400000
+                }
+            ]
+        },
+        {
+            "name": "America/Mazatlan",
+            "rules": [
+                {
+                    "e": 1414310400000,
+                    "s": 1396774800000
+                },
+                {
+                    "e": 1445760000000,
+                    "s": 1428224400000
+                },
+                {
+                    "e": 1477814400000,
+                    "s": 1459674000000
+                },
+                {
+                    "e": 1509264000000,
+                    "s": 1491123600000
+                },
+                {
+                    "e": 1540713600000,
+                    "s": 1522573200000
+                }
+            ]
+        },
+        {
+            "name": "America/Mexico_City",
+            "rules": [
+                {
+                    "e": 1414306800000,
+                    "s": 1396771200000
+                },
+                {
+                    "e": 1445756400000,
+                    "s": 1428220800000
+                },
+                {
+                    "e": 1477810800000,
+                    "s": 1459670400000
+                },
+                {
+                    "e": 1509260400000,
+                    "s": 1491120000000
+                },
+                {
+                    "e": 1540710000000,
+                    "s": 1522569600000
+                }
+            ]
+        },
+        {
+            "name": "America/Miquelon",
+            "rules": [
+                {
+                    "e": 1414900800000,
+                    "s": 1394341200000
+                },
+                {
+                    "e": 1446350400000,
+                    "s": 1425790800000
+                },
+                {
+                    "e": 1478404800000,
+                    "s": 1457845200000
+                },
+                {
+                    "e": 1509854400000,
+                    "s": 1489294800000
+                },
+                {
+                    "e": 1541304000000,
+                    "s": 1520744400000
+                }
+            ]
+        },
+        {
+            "name": "America/Santa_Isabel",
+            "rules": [
+                {
+                    "e": 1414918800000,
+                    "s": 1394359200000
+                },
+                {
+                    "e": 1446368400000,
+                    "s": 1425808800000
+                },
+                {
+                    "e": 1478422800000,
+                    "s": 1457863200000
+                },
+                {
+                    "e": 1509872400000,
+                    "s": 1489312800000
+                },
+                {
+                    "e": 1541322000000,
+                    "s": 1520762400000
+                }
+            ]
+        },
+        {
+            "name": "America/Santiago",
+            "rules": [
+                {
+                    "e": 1398567600000,
+                    "s": 1410062400000
+                },
+                false,
+                {
+                    "e": 1463281200000,
+                    "s": 1471147200000
+                },
+                {
+                    "e": 1494730800000,
+                    "s": 1502596800000
+                },
+                {
+                    "e": 1526180400000,
+                    "s": 1534046400000
+                }
+            ]
+        },
+        {
+            "name": "America/Sao_Paulo",
+            "rules": [
+                {
+                    "e": 1392516000000,
+                    "s": 1413687600000
+                },
+                {
+                    "e": 1424570400000,
+                    "s": 1445137200000
+                },
+                {
+                    "e": 1456020000000,
+                    "s": 1476586800000
+                },
+                {
+                    "e": 1487469600000,
+                    "s": 1508036400000
+                },
+                {
+                    "e": 1518919200000,
+                    "s": 1541300400000
+                }
+            ]
+        },
+        {
+            "name": "Asia/Amman",
+            "rules": [
+                {
+                    "e": 1414706400000,
+                    "s": 1395957600000
+                },
+                {
+                    "e": 1446156000000,
+                    "s": 1427407200000
+                },
+                {
+                    "e": 1477605600000,
+                    "s": 1459461600000
+                },
+                {
+                    "e": 1509055200000,
+                    "s": 1490911200000
+                },
+                {
+                    "e": 1540504800000,
+                    "s": 1522360800000
+                }
+            ]
+        },
+        {
+            "name": "Asia/Damascus",
+            "rules": [
+                {
+                    "e": 1414702800000,
+                    "s": 1395957600000
+                },
+                {
+                    "e": 1446152400000,
+                    "s": 1427407200000
+                },
+                {
+                    "e": 1477602000000,
+                    "s": 1458856800000
+                },
+                {
+                    "e": 1509051600000,
+                    "s": 1490911200000
+                },
+                {
+                    "e": 1540501200000,
+                    "s": 1522360800000
+                }
+            ]
+        },
+        {
+            "name": "Asia/Dubai",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Gaza",
+            "rules": [
+                {
+                    "e": 1414098000000,
+                    "s": 1395957600000
+                },
+                {
+                    "e": 1445547600000,
+                    "s": 1427493600000
+                },
+                {
+                    "e": 1477692000000,
+                    "s": 1458946800000
+                },
+                {
+                    "e": 1509141600000,
+                    "s": 1490396400000
+                },
+                {
+                    "e": 1540591200000,
+                    "s": 1521846000000
+                }
+            ]
+        },
+        {
+            "name": "Asia/Irkutsk",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Jerusalem",
+            "rules": [
+                {
+                    "e": 1414278000000,
+                    "s": 1395964800000
+                },
+                {
+                    "e": 1445727600000,
+                    "s": 1427414400000
+                },
+                {
+                    "e": 1477782000000,
+                    "s": 1458864000000
+                },
+                {
+                    "e": 1509231600000,
+                    "s": 1490313600000
+                },
+                {
+                    "e": 1540681200000,
+                    "s": 1521763200000
+                }
+            ]
+        },
+        {
+            "name": "Asia/Kamchatka",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Krasnoyarsk",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Omsk",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Vladivostok",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Yakutsk",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Yekaterinburg",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Asia/Yerevan",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Australia/Lord_Howe",
+            "rules": [
+                {
+                    "e": 1396710000000,
+                    "s": 1412436600000
+                },
+                {
+                    "e": 1428159600000,
+                    "s": 1443886200000
+                },
+                {
+                    "e": 1459609200000,
+                    "s": 1475335800000
+                },
+                {
+                    "e": 1491058800000,
+                    "s": 1506785400000
+                },
+                {
+                    "e": 1522508400000,
+                    "s": 1538839800000
+                }
+            ]
+        },
+        {
+            "name": "Australia/Perth",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Europe/Helsinki",
+            "rules": [
+                {
+                    "e": 1414285200000,
+                    "s": 1396141200000
+                },
+                {
+                    "e": 1445734800000,
+                    "s": 1427590800000
+                },
+                {
+                    "e": 1477789200000,
+                    "s": 1459040400000
+                },
+                {
+                    "e": 1509238800000,
+                    "s": 1490490000000
+                },
+                {
+                    "e": 1540688400000,
+                    "s": 1521939600000
+                }
+            ]
+        },
+        {
+            "name": "Europe/Minsk",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Europe/Moscow",
+            "rules": [
+                false,
+                false,
+                false,
+                false,
+                false
+            ]
+        },
+        {
+            "name": "Pacific/Apia",
+            "rules": [
+                {
+                    "e": 1396706400000,
+                    "s": 1411826400000
+                },
+                {
+                    "e": 1428156000000,
+                    "s": 1443276000000
+                },
+                {
+                    "e": 1459605600000,
+                    "s": 1474725600000
+                },
+                {
+                    "e": 1491055200000,
+                    "s": 1506175200000
+                },
+                {
+                    "e": 1522504800000,
+                    "s": 1538229600000
+                }
+            ]
+        },
+        {
+            "name": "Pacific/Fiji",
+            "rules": [
+                {
+                    "e": 1421503200000,
+                    "s": 1414850400000
+                },
+                {
+                    "e": 1452952800000,
+                    "s": 1446300000000
+                },
+                {
+                    "e": 1484402400000,
+                    "s": 1478354400000
+                },
+                {
+                    "e": 1515852000000,
+                    "s": 1509804000000
+                },
+                {
+                    "e": 1547906400000,
+                    "s": 1541253600000
+                }
+            ]
+        },
+        {
+            "name": "Europe/London",
+            "rules": [
+                {
+                    "e": 1414285200000,
+                    "s": 1396141200000
+                },
+                {
+                    "e": 1445734800000,
+                    "s": 1427590800000
+                },
+                {
+                    "e": 1477789200000,
+                    "s": 1459040400000
+                },
+                {
+                    "e": 1509238800000,
+                    "s": 1490490000000
+                },
+                {
+                    "e": 1540688400000,
+                    "s": 1521939600000
+                }
+            ]
+        }
+    ]
+};
